Fix : Quantitiy에 맞춰서 바인딩 되지 않는 현상 수정

+++ app/src/main/java/com/example/fitnutrijournal/viewmodel/HomeViewModel.kt
@@ -520,7 +520,7 @@ class HomeViewModel(application: Application) : AndroidViewModel(application) {
                     date = date,
                     mealType = mealType,
                     dietFoodCode = food.foodCd,
-                    quantity = food.servingSize.toFloat()
+                    quantity = food.servingSize
                 )
             }
 
@@ -640,7 +640,7 @@ class HomeViewModel(application: Application) : AndroidViewModel(application) {
         }
     }
 
-    fun updateNutrientData(mealType: String, food: Food, quantity: Float) {
+    fun updateNutrientData(mealType: String, food: Food, quantity: Int) {
         val nutrientData = when (mealType) {
             "breakfast" -> _breakfastNutrients.value ?: NutrientData()
             "lunch" -> _lunchNutrients.value ?: NutrientData()
